requirements.txt
pytest==8.3.0
requests
pytest-asyncio==0.23.5
playwright>=1.49.0
pytest-playwright>=0.6.0
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.5
pytest-rerunfailures==14.0
allure-python-commons
locust